Code cleanup.
